Sync React with Haste-style imports rewritten to use path-based imports instead
This commit uses the same base commit of React that's currently in RN (React ec6691a68716bc59291746fc62f374a56fb435c9) plus a commit in the React repo that removes Haste-style imports from the renderer and a commit to make the shims import from `implementations` (https://github.com/facebook/react/pull/15786).

I built React in the React repo with `yarn build` and copied over the `oss` directory into one called `implementations` and removed the `*.fb.js` files.

Test plan: run unit tests, RNTester, CI. Also verified that we can make Metro pick up `.fb.js` files with:

```js
resolver: {
  sourceExts: ['fb.js', 'js', 'json', ...],
}
```
