+++ Libraries/Renderer/REVISION
@@ -1,1 +1,1 @@
-ec6691a68716bc59291746fc62f374a56fb435c9
+87d375afd0d9394646a4bcddcebb7b15ffaa7e9e
