+++ Libraries/Renderer/implementations/ReactNativeRenderer-prod.js
@@ -11,16 +11,10 @@
  */
 
 "use strict";
-require("InitializeCore");
-var ReactNativeViewConfigRegistry = require("ReactNativeViewConfigRegistry"),
-  UIManager = require("UIManager"),
-  RCTEventEmitter = require("RCTEventEmitter"),
+require("react-native/Libraries/ReactPrivate/ReactNativePrivateInitializeCore");
+var ReactNativePrivateInterface = require("react-native/Libraries/ReactPrivate/ReactNativePrivateInterface"),
   React = require("react"),
-  deepDiffer = require("deepDiffer"),
-  flattenStyle = require("flattenStyle"),
-  TextInputState = require("TextInputState"),
   Scheduler = require("scheduler");
-var ExceptionsManager = require("ExceptionsManager");
 function ReactError(message) {
   message = Error(message);
   message.name = "Invariant Violation";
@@ -617,7 +611,7 @@ function changeResponder(nextResponderInst, blockHostResponder) {
       blockHostResponder
     );
 }
-var eventTypes$1 = {
+var eventTypes = {
     startShouldSetResponder: {
       phasedRegistrationNames: {
         bubbled: "onStartShouldSetResponder",
@@ -680,7 +674,7 @@ var eventTypes$1 = {
     _getResponder: function() {
       return responderInst;
     },
-    eventTypes: eventTypes$1,
+    eventTypes: eventTypes,
     extractEvents: function(
       topLevelType,
       targetInst,
@@ -709,12 +703,12 @@ var eventTypes$1 = {
           isMoveish(topLevelType))
       ) {
         var JSCompiler_temp = isStartish(topLevelType)
-          ? eventTypes$1.startShouldSetResponder
+          ? eventTypes.startShouldSetResponder
           : isMoveish(topLevelType)
-            ? eventTypes$1.moveShouldSetResponder
+            ? eventTypes.moveShouldSetResponder
             : "topSelectionChange" === topLevelType
-              ? eventTypes$1.selectionChangeShouldSetResponder
-              : eventTypes$1.scrollShouldSetResponder;
+              ? eventTypes.selectionChangeShouldSetResponder
+              : eventTypes.scrollShouldSetResponder;
         if (responderInst)
           b: {
             var JSCompiler_temp$jscomp$0 = responderInst;
@@ -800,7 +794,7 @@ var eventTypes$1 = {
         JSCompiler_temp && JSCompiler_temp !== responderInst
           ? ((JSCompiler_temp$jscomp$0 = void 0),
             (targetInst = ResponderSyntheticEvent.getPooled(
-              eventTypes$1.responderGrant,
+              eventTypes.responderGrant,
               JSCompiler_temp,
               nativeEvent,
               nativeEventTarget
@@ -810,7 +804,7 @@ var eventTypes$1 = {
             (depthA = !0 === executeDirectDispatch(targetInst)),
             responderInst
               ? ((tempA = ResponderSyntheticEvent.getPooled(
-                  eventTypes$1.responderTerminationRequest,
+                  eventTypes.responderTerminationRequest,
                   responderInst,
                   nativeEvent,
                   nativeEventTarget
@@ -822,7 +816,7 @@ var eventTypes$1 = {
                 tempA.isPersistent() || tempA.constructor.release(tempA),
                 tempB
                   ? ((tempA = ResponderSyntheticEvent.getPooled(
-                      eventTypes$1.responderTerminate,
+                      eventTypes.responderTerminate,
                       responderInst,
                       nativeEvent,
                       nativeEventTarget
@@ -836,7 +830,7 @@ var eventTypes$1 = {
                     )),
                     changeResponder(JSCompiler_temp, depthA))
                   : ((JSCompiler_temp = ResponderSyntheticEvent.getPooled(
-                      eventTypes$1.responderReject,
+                      eventTypes.responderReject,
                       JSCompiler_temp,
                       nativeEvent,
                       nativeEventTarget
@@ -866,11 +860,11 @@ var eventTypes$1 = {
         ("topTouchEnd" === topLevelType || "topTouchCancel" === topLevelType);
       if (
         (JSCompiler_temp$jscomp$0 = JSCompiler_temp$jscomp$0
-          ? eventTypes$1.responderStart
+          ? eventTypes.responderStart
           : targetInst
-            ? eventTypes$1.responderMove
+            ? eventTypes.responderMove
             : depthA
-              ? eventTypes$1.responderEnd
+              ? eventTypes.responderEnd
               : null)
       )
         (JSCompiler_temp$jscomp$0 = ResponderSyntheticEvent.getPooled(
@@ -924,9 +918,9 @@ var eventTypes$1 = {
         }
       if (
         (topLevelType = JSCompiler_temp$jscomp$0
-          ? eventTypes$1.responderTerminate
+          ? eventTypes.responderTerminate
           : topLevelType
-            ? eventTypes$1.responderRelease
+            ? eventTypes.responderRelease
             : null)
       )
         (nativeEvent = ResponderSyntheticEvent.getPooled(
@@ -948,8 +942,15 @@ var eventTypes$1 = {
       }
     }
   },
+  customBubblingEventTypes =
+    ReactNativePrivateInterface.ReactNativeViewConfigRegistry
+      .customBubblingEventTypes,
+  customDirectEventTypes =
+    ReactNativePrivateInterface.ReactNativeViewConfigRegistry
+      .customDirectEventTypes,
   ReactNativeBridgeEventPlugin = {
-    eventTypes: ReactNativeViewConfigRegistry.eventTypes,
+    eventTypes:
+      ReactNativePrivateInterface.ReactNativeViewConfigRegistry.eventTypes,
     extractEvents: function(
       topLevelType,
       targetInst,
@@ -957,10 +958,8 @@ var eventTypes$1 = {
       nativeEventTarget
     ) {
       if (null == targetInst) return null;
-      var bubbleDispatchConfig =
-          ReactNativeViewConfigRegistry.customBubblingEventTypes[topLevelType],
-        directDispatchConfig =
-          ReactNativeViewConfigRegistry.customDirectEventTypes[topLevelType];
+      var bubbleDispatchConfig = customBubblingEventTypes[topLevelType],
+        directDispatchConfig = customDirectEventTypes[topLevelType];
       if (!bubbleDispatchConfig && !directDispatchConfig)
         throw ReactError(
           'Unsupported top level event type "' + topLevelType + '" dispatched'
@@ -1060,7 +1059,7 @@ function _receiveRootNodeIDEvent(rootNodeID, topLevelType, nativeEventParam) {
     }
   });
 }
-RCTEventEmitter.register({
+ReactNativePrivateInterface.RCTEventEmitter.register({
   receiveEvent: function(rootNodeID, topLevelType, nativeEventParam) {
     _receiveRootNodeIDEvent(rootNodeID, topLevelType, nativeEventParam);
   },
@@ -1110,8 +1109,11 @@ getNodeFromInstance = function(inst) {
 ResponderEventPlugin.injection.injectGlobalResponderHandler({
   onChange: function(from, to, blockNativeResponder) {
     null !== to
-      ? UIManager.setJSResponder(to.stateNode._nativeTag, blockNativeResponder)
-      : UIManager.clearJSResponder();
+      ? ReactNativePrivateInterface.UIManager.setJSResponder(
+          to.stateNode._nativeTag,
+          blockNativeResponder
+        )
+      : ReactNativePrivateInterface.UIManager.clearJSResponder();
   }
 });
 var ReactSharedInternals =
@@ -1375,14 +1377,14 @@ function diffNestedProperty(
   return Array.isArray(prevProp)
     ? diffProperties(
         updatePayload,
-        flattenStyle(prevProp),
+        ReactNativePrivateInterface.flattenStyle(prevProp),
         nextProp,
         validAttributes
       )
     : diffProperties(
         updatePayload,
         prevProp,
-        flattenStyle(nextProp),
+        ReactNativePrivateInterface.flattenStyle(nextProp),
         validAttributes
       );
 }
@@ -1450,7 +1452,7 @@ function diffProperties(updatePayload, prevProps, nextProps, validAttributes) {
         if ("object" !== typeof attributeConfig)
           ("object" !== typeof nextProp ||
             null === nextProp ||
-            deepDiffer(prevProp, nextProp)) &&
+            ReactNativePrivateInterface.deepDiffer(prevProp, nextProp)) &&
             ((updatePayload || (updatePayload = {}))[propKey] = nextProp);
         else if (
           "function" === typeof attributeConfig.diff ||
@@ -1462,7 +1464,7 @@ function diffProperties(updatePayload, prevProps, nextProps, validAttributes) {
               ? attributeConfig.diff(prevProp, nextProp)
               : "object" !== typeof nextProp ||
                 null === nextProp ||
-                deepDiffer(prevProp, nextProp))
+                ReactNativePrivateInterface.deepDiffer(prevProp, nextProp))
           )
             (attributeConfig =
               "function" === typeof attributeConfig.process
@@ -1527,19 +1529,19 @@ var ReactNativeFiberHostComponent = (function() {
     this.viewConfig = viewConfig;
   }
   ReactNativeFiberHostComponent.prototype.blur = function() {
-    TextInputState.blurTextInput(this._nativeTag);
+    ReactNativePrivateInterface.TextInputState.blurTextInput(this._nativeTag);
   };
   ReactNativeFiberHostComponent.prototype.focus = function() {
-    TextInputState.focusTextInput(this._nativeTag);
+    ReactNativePrivateInterface.TextInputState.focusTextInput(this._nativeTag);
   };
   ReactNativeFiberHostComponent.prototype.measure = function(callback) {
-    UIManager.measure(
+    ReactNativePrivateInterface.UIManager.measure(
       this._nativeTag,
       mountSafeCallback_NOT_REALLY_SAFE(this, callback)
     );
   };
   ReactNativeFiberHostComponent.prototype.measureInWindow = function(callback) {
-    UIManager.measureInWindow(
+    ReactNativePrivateInterface.UIManager.measureInWindow(
       this._nativeTag,
       mountSafeCallback_NOT_REALLY_SAFE(this, callback)
     );
@@ -1558,7 +1560,7 @@ var ReactNativeFiberHostComponent = (function() {
           relativeToNativeNode.canonical._nativeTag &&
           (relativeNode = relativeToNativeNode.canonical._nativeTag);
     null != relativeNode &&
-      UIManager.measureLayout(
+      ReactNativePrivateInterface.UIManager.measureLayout(
         this._nativeTag,
         relativeNode,
         mountSafeCallback_NOT_REALLY_SAFE(this, onFail),
@@ -1575,7 +1577,7 @@ var ReactNativeFiberHostComponent = (function() {
       this.viewConfig.validAttributes
     );
     null != nativeProps &&
-      UIManager.updateView(
+      ReactNativePrivateInterface.UIManager.updateView(
         this._nativeTag,
         this.viewConfig.uiViewClassName,
         nativeProps
@@ -1588,7 +1590,9 @@ function shim$1() {
     "The current renderer does not support hydration. This error is likely caused by a bug in React. Please file an issue."
   );
 }
-var UPDATE_SIGNAL = {},
+var getViewConfigForType =
+    ReactNativePrivateInterface.ReactNativeViewConfigRegistry.get,
+  UPDATE_SIGNAL = {},
   nextReactTag = 3;
 function allocateTag() {
   var tag = nextReactTag;
@@ -1611,7 +1615,10 @@ function finalizeInitialChildren(parentInstance) {
   var nativeTags = parentInstance._children.map(function(child) {
     return "number" === typeof child ? child : child._nativeTag;
   });
-  UIManager.setChildren(parentInstance._nativeTag, nativeTags);
+  ReactNativePrivateInterface.UIManager.setChildren(
+    parentInstance._nativeTag,
+    nativeTags
+  );
   return !1;
 }
 var scheduleTimeout = setTimeout,
@@ -4535,7 +4542,7 @@ function logCapturedError(capturedError) {
       "string" === typeof error
         ? Error(error + "\n\nThis error is located at:" + componentStack)
         : Error("Unspecified error at:" + componentStack);
-  ExceptionsManager.handleException(error, !1);
+  ReactNativePrivateInterface.ExceptionsManager.handleException(error, !1);
 }
 var PossiblyWeakSet$1 = "function" === typeof WeakSet ? WeakSet : Set;
 function logError(boundary, errorInfo) {
@@ -4607,7 +4614,7 @@ function hideOrUnhideAllChildren(finishedWork, isHidden) {
           { style: { display: "none" } },
           viewConfig.validAttributes
         );
-        UIManager.updateView(
+        ReactNativePrivateInterface.UIManager.updateView(
           instance._nativeTag,
           viewConfig.uiViewClassName,
           updatePayload
@@ -4625,7 +4632,7 @@ function hideOrUnhideAllChildren(finishedWork, isHidden) {
           updatePayload,
           viewConfig.validAttributes
         );
-        UIManager.updateView(
+        ReactNativePrivateInterface.UIManager.updateView(
           instance._nativeTag,
           viewConfig.uiViewClassName,
           updatePayload
@@ -4778,7 +4785,7 @@ function commitPlacement(finishedWork) {
             ? (children.splice(index, 1),
               (beforeChild = children.indexOf(beforeChild)),
               children.splice(beforeChild, 0, stateNode),
-              UIManager.manageChildren(
+              ReactNativePrivateInterface.UIManager.manageChildren(
                 parentInstance._nativeTag,
                 [index],
                 [beforeChild],
@@ -4788,7 +4795,7 @@ function commitPlacement(finishedWork) {
               ))
             : ((index = children.indexOf(beforeChild)),
               children.splice(index, 0, stateNode),
-              UIManager.manageChildren(
+              ReactNativePrivateInterface.UIManager.manageChildren(
                 parentInstance._nativeTag,
                 [],
                 [],
@@ -4803,7 +4810,7 @@ function commitPlacement(finishedWork) {
         }
       else
         isContainer
-          ? UIManager.setChildren(parent, [
+          ? ReactNativePrivateInterface.UIManager.setChildren(parent, [
               "number" === typeof stateNode ? stateNode : stateNode._nativeTag
             ])
           : ((parentInstance = parent),
@@ -4814,7 +4821,7 @@ function commitPlacement(finishedWork) {
             0 <= beforeChild
               ? (index.splice(beforeChild, 1),
                 index.push(stateNode),
-                UIManager.manageChildren(
+                ReactNativePrivateInterface.UIManager.manageChildren(
                   parentInstance._nativeTag,
                   [beforeChild],
                   [index.length - 1],
@@ -4823,7 +4830,7 @@ function commitPlacement(finishedWork) {
                   []
                 ))
               : (index.push(stateNode),
-                UIManager.manageChildren(
+                ReactNativePrivateInterface.UIManager.manageChildren(
                   parentInstance._nativeTag,
                   [],
                   [],
@@ -4900,7 +4907,14 @@ function unmountHostComponents(current$$1) {
       if (currentParentIsContainer)
         (root = currentParent),
           recursivelyUncacheFiberNode(node.stateNode),
-          UIManager.manageChildren(root, [], [], [], [], [0]);
+          ReactNativePrivateInterface.UIManager.manageChildren(
+            root,
+            [],
+            [],
+            [],
+            [],
+            [0]
+          );
       else {
         root = currentParent;
         var child = node.stateNode;
@@ -4908,7 +4922,14 @@ function unmountHostComponents(current$$1) {
         node$jscomp$0 = root._children;
         child = node$jscomp$0.indexOf(child);
         node$jscomp$0.splice(child, 1);
-        UIManager.manageChildren(root._nativeTag, [], [], [], [], [child]);
+        ReactNativePrivateInterface.UIManager.manageChildren(
+          root._nativeTag,
+          [],
+          [],
+          [],
+          [],
+          [child]
+        );
       }
     } else if (4 === node.tag) {
       if (null !== node.child) {
@@ -4960,7 +4981,7 @@ function commitWork(current$$1, finishedWork) {
             finishedWork.validAttributes
           )),
           null != newProps &&
-            UIManager.updateView(
+            ReactNativePrivateInterface.UIManager.updateView(
               instance._nativeTag,
               finishedWork.uiViewClassName,
               newProps
@@ -4972,9 +4993,11 @@ function commitWork(current$$1, finishedWork) {
         throw ReactError(
           "This should have a text node initialized. This error is likely caused by a bug in React. Please file an issue."
         );
-      UIManager.updateView(finishedWork.stateNode, "RCTRawText", {
-        text: finishedWork.memoizedProps
-      });
+      ReactNativePrivateInterface.UIManager.updateView(
+        finishedWork.stateNode,
+        "RCTRawText",
+        { text: finishedWork.memoizedProps }
+      );
       break;
     case 20:
       break;
@@ -5643,14 +5666,14 @@ function completeUnitOfWork(unitOfWork) {
               var rootContainerInstance = renderExpirationTime$jscomp$0,
                 internalInstanceHandle = current$$1,
                 tag = allocateTag();
-              type$jscomp$0 = ReactNativeViewConfigRegistry.get(type$jscomp$0);
+              type$jscomp$0 = getViewConfigForType(type$jscomp$0);
               var updatePayload = diffProperties(
                 null,
                 emptyObject,
                 instance,
                 type$jscomp$0.validAttributes
               );
-              UIManager.createView(
+              ReactNativePrivateInterface.UIManager.createView(
                 tag,
                 type$jscomp$0.uiViewClassName,
                 rootContainerInstance,
@@ -5699,9 +5722,12 @@ function completeUnitOfWork(unitOfWork) {
                   "Text strings must be rendered within a <Text> component."
                 );
               type = allocateTag();
-              UIManager.createView(type, "RCTRawText", current, {
-                text: newProps
-              });
+              ReactNativePrivateInterface.UIManager.createView(
+                type,
+                "RCTRawText",
+                current,
+                { text: newProps }
+              );
               instanceCache[type] = current$$1;
               renderExpirationTime$jscomp$0.stateNode = type;
             }
@@ -6899,10 +6925,14 @@ var roots = new Map(),
         }
         _inherits(ReactNativeComponent, _React$Component);
         ReactNativeComponent.prototype.blur = function() {
-          TextInputState.blurTextInput(findNodeHandle(this));
+          ReactNativePrivateInterface.TextInputState.blurTextInput(
+            findNodeHandle(this)
+          );
         };
         ReactNativeComponent.prototype.focus = function() {
-          TextInputState.focusTextInput(findNodeHandle(this));
+          ReactNativePrivateInterface.TextInputState.focusTextInput(
+            findNodeHandle(this)
+          );
         };
         ReactNativeComponent.prototype.measure = function(callback) {
           var maybeInstance = void 0;
@@ -6915,7 +6945,7 @@ var roots = new Map(),
                   maybeInstance.node,
                   mountSafeCallback_NOT_REALLY_SAFE(this, callback)
                 )
-              : UIManager.measure(
+              : ReactNativePrivateInterface.UIManager.measure(
                   findNodeHandle(this),
                   mountSafeCallback_NOT_REALLY_SAFE(this, callback)
                 ));
@@ -6931,7 +6961,7 @@ var roots = new Map(),
                   maybeInstance.node,
                   mountSafeCallback_NOT_REALLY_SAFE(this, callback)
                 )
-              : UIManager.measureInWindow(
+              : ReactNativePrivateInterface.UIManager.measureInWindow(
                   findNodeHandle(this),
                   mountSafeCallback_NOT_REALLY_SAFE(this, callback)
                 ));
@@ -6953,7 +6983,7 @@ var roots = new Map(),
               : relativeToNativeNode._nativeTag &&
                 (maybeInstance = relativeToNativeNode._nativeTag),
             null != maybeInstance &&
-              UIManager.measureLayout(
+              ReactNativePrivateInterface.UIManager.measureLayout(
                 findNodeHandle(this),
                 maybeInstance,
                 mountSafeCallback_NOT_REALLY_SAFE(this, onFail),
@@ -6977,7 +7007,7 @@ var roots = new Map(),
               maybeInstance.validAttributes
             );
             null != nativeProps &&
-              UIManager.updateView(
+              ReactNativePrivateInterface.UIManager.updateView(
                 nativeTag,
                 maybeInstance.uiViewClassName,
                 nativeProps
@@ -6997,7 +7027,7 @@ var roots = new Map(),
           handle.viewConfig.validAttributes
         )),
         null != nativeProps &&
-          UIManager.updateView(
+          ReactNativePrivateInterface.UIManager.updateView(
             handle._nativeTag,
             handle.viewConfig.uiViewClassName,
             nativeProps
@@ -7033,7 +7063,7 @@ var roots = new Map(),
     },
     unmountComponentAtNodeAndRemoveContainer: function(containerTag) {
       ReactNativeRenderer.unmountComponentAtNode(containerTag);
-      UIManager.removeRootView(containerTag);
+      ReactNativePrivateInterface.UIManager.removeRootView(containerTag);
     },
     createPortal: function(children, containerTag) {
       return createPortal(
@@ -7058,7 +7088,7 @@ var roots = new Map(),
                     maybeInstance.node,
                     mountSafeCallback_NOT_REALLY_SAFE(this, callback)
                   )
-                : UIManager.measure(
+                : ReactNativePrivateInterface.UIManager.measure(
                     findNodeHandle(this),
                     mountSafeCallback_NOT_REALLY_SAFE(this, callback)
                   ));
@@ -7074,7 +7104,7 @@ var roots = new Map(),
                     maybeInstance.node,
                     mountSafeCallback_NOT_REALLY_SAFE(this, callback)
                   )
-                : UIManager.measureInWindow(
+                : ReactNativePrivateInterface.UIManager.measureInWindow(
                     findNodeHandle(this),
                     mountSafeCallback_NOT_REALLY_SAFE(this, callback)
                   ));
@@ -7092,7 +7122,7 @@ var roots = new Map(),
                 : relativeToNativeNode._nativeTag &&
                   (maybeInstance = relativeToNativeNode._nativeTag),
               null != maybeInstance &&
-                UIManager.measureLayout(
+                ReactNativePrivateInterface.UIManager.measureLayout(
                   findNodeHandle(this),
                   maybeInstance,
                   mountSafeCallback_NOT_REALLY_SAFE(this, onFail),
@@ -7116,7 +7146,7 @@ var roots = new Map(),
                 maybeInstance.validAttributes
               );
               null != nativeProps &&
-                UIManager.updateView(
+                ReactNativePrivateInterface.UIManager.updateView(
                   nativeTag,
                   maybeInstance.uiViewClassName,
                   nativeProps
@@ -7124,10 +7154,14 @@ var roots = new Map(),
             }
           },
           focus: function() {
-            TextInputState.focusTextInput(findNodeHandle(this));
+            ReactNativePrivateInterface.TextInputState.focusTextInput(
+              findNodeHandle(this)
+            );
           },
           blur: function() {
-            TextInputState.blurTextInput(findNodeHandle(this));
+            ReactNativePrivateInterface.TextInputState.blurTextInput(
+              findNodeHandle(this)
+            );
           }
         };
       })(findNodeHandle, findHostInstance),
