+++ Libraries/Renderer/shims/ReactFabric.js
@@ -10,7 +10,7 @@
 
 'use strict';
 
-const BatchedBridge = require('../../BatchedBridge/BatchedBridge');
+import {BatchedBridge} from 'react-native/Libraries/ReactPrivate/ReactNativePrivateInterface';
 
 // TODO @sema: Adjust types
 import type {ReactNativeType} from './ReactNativeTypes';
@@ -18,9 +18,9 @@ import type {ReactNativeType} from './ReactNativeTypes';
 let ReactFabric;
 
 if (__DEV__) {
-  ReactFabric = require('ReactFabric-dev');
+  ReactFabric = require('../implementations/ReactFabric-dev');
 } else {
-  ReactFabric = require('ReactFabric-prod');
+  ReactFabric = require('../implementations/ReactFabric-prod');
 }
 
 BatchedBridge.registerCallableModule('ReactFabric', ReactFabric);
